Maven Market Power Bi Dashboard
Created this dashboard to track the transaction of the grocery store. 
Transaction is tracked by store cities and countries to analyze the large consumer base areas.
Transaction is also shown by the start of the week to analyze the trend week by week.
Revenue is compared against the target to see if the year was closed in green.